garbage collector trait added
