+++ defuse/src/contract/accounts/mod.rs
@@ -7,7 +7,7 @@ pub use self::{account::*, state::*};
 use std::{borrow::Cow, collections::HashSet};
 
 use defuse_core::{
-    DefuseError, ExpirableNonce, Nonce, SaltedNonce, VersionedNonce,
+    DefuseError, Nonce,
     accounts::{AccountEvent, PublicKeyEvent},
     crypto::PublicKey,
     engine::{State, StateView},
@@ -16,7 +16,6 @@ use defuse_core::{
 use defuse_near_utils::{Lock, NestPrefix, PREDECESSOR_ACCOUNT_ID, UnwrapOrPanic};
 use defuse_serde_utils::base64::AsBase64;
 
-use near_plugins::{AccessControllable, access_control_any};
 use near_sdk::{
     AccountId, AccountIdRef, BorshStorageKey, FunctionError, IntoStorageKey, assert_one_yocto,
     borsh::BorshSerialize, near, store::IterableMap,
@@ -24,7 +23,7 @@ use near_sdk::{
 
 use crate::{
     accounts::AccountManager,
-    contract::{Contract, ContractExt, Role, accounts::AccountEntry},
+    contract::{Contract, ContractExt, accounts::AccountEntry},
 };
 
 #[near]
@@ -71,24 +70,6 @@ impl AccountManager for Contract {
         StateView::is_nonce_used(self, account_id, nonce.into_inner())
     }
 
-    #[access_control_any(roles(Role::DAO, Role::GarbageCollector))]
-    #[payable]
-    fn cleanup_nonces(&mut self, nonces: Vec<(AccountId, Vec<AsBase64<Nonce>>)>) {
-        assert_one_yocto();
-
-        for (account_id, nonces) in nonces {
-            for nonce in nonces.into_iter().map(AsBase64::into_inner) {
-                if !self.is_nonce_cleanable(nonce) {
-                    continue;
-                }
-
-                // NOTE: all errors are omitted
-                let [prefix @ .., _] = nonce;
-                let _ = State::cleanup_nonce_by_prefix(self, &account_id, prefix);
-            }
-        }
-    }
-
     fn is_auth_by_predecessor_id_enabled(&self, account_id: &AccountId) -> bool {
         StateView::is_auth_by_predecessor_id_enabled(self, account_id)
     }
@@ -102,20 +83,6 @@ impl AccountManager for Contract {
 }
 
 impl Contract {
-    #[inline]
-    fn is_nonce_cleanable(&self, nonce: Nonce) -> bool {
-        let Some(versioned_nonce) = VersionedNonce::maybe_from(nonce) else {
-            return false;
-        };
-
-        match versioned_nonce {
-            VersionedNonce::V1(SaltedNonce {
-                salt,
-                nonce: ExpirableNonce { deadline, .. },
-            }) => deadline.has_expired() || !self.is_valid_salt(salt),
-        }
-    }
-
     #[inline]
     pub fn ensure_auth_predecessor_id(&self) -> &'static AccountId {
         if !StateView::is_auth_by_predecessor_id_enabled(self, &PREDECESSOR_ACCOUNT_ID) {
