+++ defuse/src/lib.rs
@@ -4,6 +4,7 @@ pub mod contract;
 
 pub mod accounts;
 pub mod fees;
+pub mod garbage_collector;
 pub mod intents;
 pub mod salts;
 pub mod simulation_output;
