+++ tests/src/tests/defuse/accounts/mod.rs
@@ -9,8 +9,6 @@ use defuse_serde_utils::base64::AsBase64;
 use near_sdk::{AccountId, AccountIdRef, Gas, NearToken};
 use serde_json::json;
 
-use crate::utils::test_log::TestLog;
-
 pub trait AccountManagerExt {
     async fn add_public_key(
         &self,
@@ -23,12 +21,6 @@ pub trait AccountManagerExt {
         defuse_contract_id: &AccountId,
         public_key: PublicKey,
     ) -> anyhow::Result<()>;
-
-    async fn cleanup_nonces(
-        &self,
-        defuse_contract_id: &AccountId,
-        data: &[(AccountId, Vec<Nonce>)],
-    ) -> anyhow::Result<TestLog>;
 
     async fn defuse_has_public_key(
         &self,
@@ -89,35 +81,6 @@ impl AccountManagerExt for near_workspaces::Account {
             .await?
             .into_result()?;
         Ok(())
-    }
-
-    async fn cleanup_nonces(
-        &self,
-        defuse_contract_id: &AccountId,
-        data: &[(AccountId, Vec<Nonce>)],
-    ) -> anyhow::Result<TestLog> {
-        let nonces = data
-            .iter()
-            .map(|(acc, nonces)| {
-                let base64_nonces: Vec<AsBase64<Nonce>> =
-                    nonces.iter().map(|nonce| AsBase64(*nonce)).collect();
-                (acc.clone(), base64_nonces)
-            })
-            .collect::<Vec<(AccountId, Vec<AsBase64<Nonce>>)>>();
-
-        let res = self
-            .call(defuse_contract_id, "cleanup_nonces")
-            .deposit(NearToken::from_yoctonear(1))
-            .args_json(json!({
-                "nonces": nonces,
-            }))
-            .max_gas()
-            .transact()
-            .await?
-            .into_result()
-            .map(TestLog::from)?;
-
-        Ok(res)
     }
 
     async fn defuse_has_public_key(
@@ -201,16 +164,6 @@ impl AccountManagerExt for near_workspaces::Contract {
     ) -> anyhow::Result<()> {
         self.as_account()
             .remove_public_key(defuse_contract_id, public_key)
-            .await
-    }
-
-    async fn cleanup_nonces(
-        &self,
-        defuse_contract_id: &AccountId,
-        data: &[(AccountId, Vec<Nonce>)],
-    ) -> anyhow::Result<TestLog> {
-        self.as_account()
-            .cleanup_nonces(defuse_contract_id, data)
             .await
     }
