+++ tests/src/tests/defuse/mod.rs
@@ -1,5 +1,6 @@
 pub mod accounts;
 mod env;
+mod garbage_collector;
 mod intents;
 mod state;
 mod storage;
